requirements.txt
streamlit
langchain
langchain-community
openai
torch
transformers
pillow
python-dotenv
watchdog